+++ weaver/processes/wps_package.py
@@ -74,6 +74,16 @@ from weaver.formats import (
 from weaver.processes import opensearch
 from weaver.processes.constants import (
     CWL_NAMESPACE_CWLTOOL_URL,
+    CWL_NAMESPACE_SCHEMA_ID,
+    CWL_NAMESPACE_SCHEMA_METADATA_AUTHOR,
+    CWL_NAMESPACE_SCHEMA_METADATA_CODE_REPOSITORY,
+    CWL_NAMESPACE_SCHEMA_METADATA_CONTRIBUTOR,
+    CWL_NAMESPACE_SCHEMA_METADATA_KEYWORDS,
+    CWL_NAMESPACE_SCHEMA_METADATA_PERSON,
+    CWL_NAMESPACE_SCHEMA_METADATA_SOFTWARE_VERSION,
+    CWL_NAMESPACE_SCHEMA_METADATA_SUPPORTED,
+    CWL_NAMESPACE_SCHEMA_METADATA_VERSION,
+    CWL_NAMESPACE_SCHEMA_URL,
     CWL_NAMESPACE_WEAVER_DEFINITION,
     CWL_NAMESPACE_WEAVER_ID,
     CWL_REQUIREMENT_APP_BUILTIN,
@@ -226,16 +236,6 @@ PACKAGE_PROGRESS_PREP_OUT = 98
 PACKAGE_PROGRESS_DONE = 100
 
 PACKAGE_SCHEMA_CACHE = {}  # type: Dict[str, Tuple[str, str]]
-
-SUPPORTED_METADATA_MAPPING = [
-    "s:author",
-    "s:citation",
-    "s:codeRepository",
-    "s:contributor",
-    "s:dateCreated",
-    "s:license",
-    "s:releaseNotes",
-]
 
 
 def get_status_location_log_path(status_location, out_dir=None):
@@ -772,81 +772,124 @@ def _get_package_inputs_outputs(package_factory,    # type: CWLFactoryCallable
             _get_package_io(package_factory, io_select=IO_OUTPUT, as_json=as_json))
 
 
-def _update_package_metadata(wps_package_metadata, cwl_package_package):
+def _update_package_metadata(wps_metadata, cwl_package):
     # type: (JSON, CWL) -> None
     """
     Updates the package :term:`WPS` metadata dictionary from extractable `CWL` package definition.
     """
-    wps_package_metadata["title"] = wps_package_metadata.get("title", cwl_package_package.get("label", ""))
-    wps_package_metadata["abstract"] = wps_package_metadata.get("abstract", cwl_package_package.get("doc", ""))
+    wps_metadata["title"] = wps_metadata.get("title", cwl_package.get("label", ""))
+    wps_metadata["abstract"] = wps_metadata.get("abstract", cwl_package.get("doc", ""))
 
     if (
-        "$schemas" in cwl_package_package
-        and isinstance(cwl_package_package["$schemas"], list)
-        and "$namespaces" in cwl_package_package
-        and isinstance(cwl_package_package["$namespaces"], dict)
+        CWL_NAMESPACE_SCHEMA_METADATA_KEYWORDS in cwl_package and
+        isinstance(cwl_package[CWL_NAMESPACE_SCHEMA_METADATA_KEYWORDS], list)
     ):
-        metadata = wps_package_metadata.get("metadata", [])
-        namespaces_inv = {v: k for k, v in cwl_package_package["$namespaces"]}
-        for schema in cwl_package_package["$schemas"]:
-            for namespace_url in namespaces_inv:
-                if schema.startswith(namespace_url):
-                    metadata.append({"title": namespaces_inv[namespace_url], "href": schema})
-        wps_package_metadata["metadata"] = metadata
-
-    if "s:keywords" in cwl_package_package and isinstance(cwl_package_package["s:keywords"], list):
-        wps_package_metadata["keywords"] = list(
-            set(wps_package_metadata.get("keywords", [])) | set(cwl_package_package.get("s:keywords", []))
+        wps_metadata["keywords"] = list(
+            set(wps_metadata.get("keywords", [])) |
+            set(cwl_package.get(CWL_NAMESPACE_SCHEMA_METADATA_KEYWORDS, []))
         )
 
     # specific use case with a different mapping
     # https://docs.ogc.org/bp/20-089r1.html#toc31
-    if "s:version" in cwl_package_package or "s:softwareVersion" in cwl_package_package:
+    if (
+        CWL_NAMESPACE_SCHEMA_METADATA_VERSION in cwl_package or
+        CWL_NAMESPACE_SCHEMA_METADATA_SOFTWARE_VERSION in cwl_package
+    ):
         version_value = (
-            wps_package_metadata.get("version")
-            or cwl_package_package.get("s:version")
-            or cwl_package_package.get("s:softwareVersion")
+            wps_metadata.get("version")
+            or cwl_package.get(CWL_NAMESPACE_SCHEMA_METADATA_VERSION)
+            or cwl_package.get(CWL_NAMESPACE_SCHEMA_METADATA_SOFTWARE_VERSION)
         )
         # Only set the key if version_value is not empty or null
         if version_value:
-            wps_package_metadata["version"] = str(version_value)
+            wps_metadata["version"] = str(version_value)
     else:
-        version_value = wps_package_metadata.get("version")
+        version_value = wps_metadata.get("version")
         if version_value:
-            wps_package_metadata["version"] = str(version_value)
+            wps_metadata["version"] = str(version_value)
 
-    for metadata_mapping in SUPPORTED_METADATA_MAPPING:
-        if metadata_mapping in cwl_package_package:
-            metadata = wps_package_metadata.get("metadata", [])
+    schema_ns = f"{CWL_NAMESPACE_SCHEMA_ID}:"
+    metadata = wps_metadata.get("metadata", [])
+    for meta_name in CWL_NAMESPACE_SCHEMA_METADATA_SUPPORTED:
+        if meta_name in [  # skip handled above
+            CWL_NAMESPACE_SCHEMA_METADATA_KEYWORDS,
+            CWL_NAMESPACE_SCHEMA_METADATA_VERSION,
+            CWL_NAMESPACE_SCHEMA_METADATA_SOFTWARE_VERSION,
+        ]:
+            continue
+
+        meta_uri = meta_name.replace(schema_ns, CWL_NAMESPACE_SCHEMA_URL)
+
+        # CWL package => WPS context
+        if meta_name in cwl_package:
             if (
-                isinstance((cwl_package_package[metadata_mapping]), str)
-                and urlparse(cwl_package_package[metadata_mapping]).scheme != ""
+                isinstance(cwl_package[meta_name], str)
+                and urlparse(cwl_package[meta_name]).scheme != ""
             ):
-                url = cwl_package_package[metadata_mapping]
-                if metadata_mapping == "s:codeRepository":
-                    type = "text/html"
+                url = cwl_package[meta_name]
+                if meta_name == CWL_NAMESPACE_SCHEMA_METADATA_CODE_REPOSITORY:
+                    ctype = ContentType.TEXT_HTML
                 else:
-                    type = get_content_type(os.path.splitext(url)[-1], default=ContentType.TEXT_PLAIN)
+                    ctype = get_content_type(os.path.splitext(url)[-1], default=ContentType.TEXT_PLAIN)
                 metadata.append({
-                    "type": type,
-                    "rel": metadata_mapping.strip("s:"),
-                    "href": cwl_package_package[metadata_mapping]
+                    "type": ctype,
+                    "rel": meta_uri,
+                    "href": cwl_package[meta_name]
+                })
+            elif isinstance(cwl_package[meta_name], str):
+                metadata.append({
+                    "role": meta_uri,
+                    "value": cwl_package[meta_name]
                 })
             else:
-                for objects in cwl_package_package[metadata_mapping]:
-                    class_name = objects["class"].strip("s:")
+                for objects in cwl_package[meta_name]:
+                    class_name = objects["class"].strip(schema_ns)
                     value = {
-                        "$schema": f"https://schema.org/{class_name}"
+                        "$schema": f"{CWL_NAMESPACE_SCHEMA_URL}{class_name}"
                     }
                     for key, val in objects.items():
-                        if key.startswith("s:"):
-                            value[key.strip("s:")] = val
+                        if key.startswith(schema_ns):
+                            value[key.strip(schema_ns)] = val
                     metadata.append({
-                        "role": metadata_mapping.strip("s:"),
+                        "role": meta_uri,
                         "value": value
                     })
+            wps_metadata["metadata"] = metadata
 
-            wps_package_metadata["metadata"] = metadata
+        # CWL package <= WPS context
+        # note:
+        #   this mapping is accomplished only if the CWL did not already result in creating the other mapping
+        #   purposely avoid overriding a field already provided in CWL (the "truth"), even if they mismatch
+        #   this could be done on purpose, such as attributing different authors for the process vs package
+        else:
+            metadata_found = [meta for meta in metadata if (meta.get("role") or meta.get("rel")) == meta_uri]
+            if not metadata_found:
+                continue
+            metadata_found = copy.deepcopy(metadata_found)
+            if meta_name in [CWL_NAMESPACE_SCHEMA_METADATA_AUTHOR, CWL_NAMESPACE_SCHEMA_METADATA_CONTRIBUTOR]:
+                for meta in metadata_found:
+                    if "value" not in meta or not isinstance(meta["value"], dict):
+                        continue  # pragma: no cover  # sanity check, should not happen (fails validation)
+                    meta_schema = meta.get("$schema", "").replace(CWL_NAMESPACE_SCHEMA_URL, schema_ns)
+                    meta_schema = meta_schema or CWL_NAMESPACE_SCHEMA_METADATA_PERSON
+                    meta["class"] = meta_schema
+                    meta.pop("role", None)
+                    meta.pop("$schema", None)
+                    meta["value"].pop("$schema", None)
+                    meta_value = meta.pop("value")
+                    for field in list(meta_value):
+                        field_ns = f"{CWL_NAMESPACE_SCHEMA_ID}:{field}"
+                        meta[field_ns] = meta_value.pop(field)
+                cwl_package[meta_name] = metadata_found
+
+            # all other fields must be a single string
+            # ignore others to avoid injecting unknown structures that could break the CWL
+            elif len(metadata_found) == 1:
+                meta_key = get_any_value(metadata_found[0], key=True)
+                meta_value = metadata_found[0][meta_key]
+                if not isinstance(meta_value, str):
+                    continue  # pragma: no cover  # sanity check, should not happen (fails validation)
+                cwl_package[meta_name] = meta_value
 
 
 def _patch_wps_process_description_url(reference, process_hint):
@@ -1896,8 +1939,8 @@ class WpsPackage(Process):
                 level=logging.WARNING if (app_euid == "0" or app_egid == "0") else logging.INFO,
             )
 
-    def update_status(self, message, progress, status, error=None):
-        # type: (str, Number, AnyStatusType, Optional[Exception]) -> None
+    def update_status(self, message, progress, status, error=None, step=False):
+        # type: (str, Number, AnyStatusType, Optional[Exception], bool) -> None
         """
         Updates the :mod:`pywps` real job status from a specified parameters.
         """
@@ -1918,7 +1961,7 @@ class WpsPackage(Process):
             # therefore, use the '_update_status' to enforce the status
             # using protected method also avoids weird overrides of progress
             # percent on failure and final 'success' status
-            self.response._update_status(pywps_status_id, message, self.percent)  # noqa: W0212
+            self.response._update_status(pywps_status_id, message, self.percent, clean=not step)  # noqa: W0212
 
         if isinstance(error, Exception):
             self.exception_message(exception_type=type(error), exception=error,
@@ -1936,11 +1979,17 @@ class WpsPackage(Process):
                            status,                  # type: AnyStatusType
                            error=None,              # type: Optional[Exception]
                            ):                       # type: (...) -> None
+        # ensure the status of the current workflow is not changed to 'success' if up a step-update
+        # setting to 'success' will report the wrong value in the status XML document
+        # this would also trigger cleanup, which would remove an staged file needed by a future step
+        if status == Status.SUCCEEDED:
+            status = Status.RUNNING
         self.update_status(
             message=f"[provider: {target_host}, step: {step_name}] - {str(message).strip()}",
             progress=map_progress(progress, start_step_progress, end_step_progress),
             status=status,
             error=error,
+            step=True,
         )
 
     def log(self, level, message, *args, **kwargs):